pyproject.toml
[project]
name = "clif-meds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clifpy>=0.3.9",
    "pyyaml>=6.0",
    "tqdm>=4.0",
]
